Finished Makefile

+++ src/Units/Unit.cpp
@@ -5,7 +5,7 @@ Unit::~Unit(){};
 void Unit::tryToRecruit(STATE *s){
   bool onValidLocation = false;
 
-  for(int i = 0; i < this->recruitableLocations.size(); i++){
+  for(int i = 0; i < (int) this->recruitableLocations.size(); i++){
     if(this->recruitableLocations[i] == s->curHex->location)
       onValidLocation = true;
   }
@@ -13,10 +13,10 @@ void Unit::tryToRecruit(STATE *s){
     return;
 
   int i = 0;
-  while(s->UnitOffer[i] != this && i <  s->UnitOffer.size())
+  while(s->UnitOffer[i] != this && i < (int) s->UnitOffer.size())
     i++;
 
-  if(i < s->UnitOffer.size()){
+  if(i < (int) s->UnitOffer.size()){
     s->UnitOffer.erase(s->UnitOffer.begin() + i);
     s->PlayerUnits.push_back(this);
     s->avInfluence = s->avInfluence - this->influenceCost;
